Moved all fragments + activity to presentation package
