+++ app/src/main/java/com/example/f1calendarOP/RaceListViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.f1calendarOP.presentation.TAG
 import kotlinx.coroutines.launch
 import retrofit2.HttpException
 import java.io.IOException
