+++ app/src/main/java/com/example/f1calendarOP/presentation/DriversFragment.kt
@@ -1,6 +1,7 @@
-package com.example.f1calendarOP
+package com.example.f1calendarOP.presentation
 
 import androidx.fragment.app.Fragment
+import com.example.f1calendarOP.R
 
 class DriversFragment : Fragment(R.layout.fragment_drivers) {
 }
