+++ app/src/main/java/com/example/f1calendarOP/presentation/MainActivity.kt
@@ -1,11 +1,11 @@
-package com.example.f1calendarOP
+package com.example.f1calendarOP.presentation
 
-import android.content.res.Configuration
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import com.example.f1calendarOP.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
