+++ app/src/main/res/navigation/nav_graph_application.xml
@@ -6,18 +6,18 @@
 
     <fragment
         android:id="@+id/constructorsFragment"
-        android:name="com.example.f1calendarOP.ConstructorsFragment"
+        android:name="com.example.f1calendarOP.presentation.ConstructorsFragment"
         android:label="ConstructorsFragment" />
     <fragment
         android:id="@+id/standingsFragment"
-        android:name="com.example.f1calendarOP.StandingsFragment"
+        android:name="com.example.f1calendarOP.presentation.StandingsFragment"
         android:label="StandingsFragment" />
     <navigation
         android:id="@+id/navigation2"
         app:startDestination="@id/raceListFragment">
         <fragment
             android:id="@+id/raceDetailFragment"
-            android:name="com.example.f1calendarOP.RaceDetailFragment"
+            android:name="com.example.f1calendarOP.presentation.RaceDetailFragment"
             android:label="DetailFragment" >
             <argument
                 android:name="raceInfo"
@@ -25,7 +25,7 @@
         </fragment>
         <fragment
             android:id="@+id/raceListFragment"
-            android:name="com.example.f1calendarOP.RaceListFragment"
+            android:name="com.example.f1calendarOP.presentation.RaceListFragment"
             android:label="ListFragment">
             <action
                 android:id="@+id/action_raceListFragment_to_raceDetailFragment"
@@ -34,6 +34,6 @@
     </navigation>
     <fragment
         android:id="@+id/driversFragment"
-        android:name="com.example.f1calendarOP.DriversFragment"
+        android:name="com.example.f1calendarOP.presentation.DriversFragment"
         android:label="DriversFragment" />
 </navigation>
